FIx PostFoem conponent
