+++ src/app/admin/posts/_components/PostForm.tsx
@@ -1,9 +1,11 @@
 'use client'
 import { Controller, Control } from "react-hook-form"
+import { useState,useEffect } from "react"
 import Select from "react-select"
 import { FieldErrors, UseFormHandleSubmit, UseFormRegister,  } from "react-hook-form"
 import { CreatePostRequestBody } from "@/app/_types/type"
 import { SelectOptionForCategories } from "@/app/_types/type"
+import { CategoryData } from "@/app/_types/type";
 
 type FormValues = {
   title: string,
@@ -22,7 +24,6 @@ interface Props {
   errors: FieldErrors<FormValues>;
   submitFunction: (data: CreatePostRequestBody) => Promise<void>;
   control: Control<CreatePostRequestBody, any>;
-  categoryOptions: SelectOptionForCategories[]
 }
 
 const PostForm: React.FC<Props> = ({ 
@@ -32,8 +33,47 @@ const PostForm: React.FC<Props> = ({
   errors, 
   submitFunction,
   control,
-  categoryOptions
 }) => {
+
+  const [ isLoading, setIsLoading ] = useState(true);
+  const [ categoryOptions, setCategoryOptions ] = useState<SelectOptionForCategories[]>([]);
+
+  // GET: カテゴリー一覧の取得
+    useEffect(() => {
+      const fetcher = async () => {
+        try {
+          const res = await fetch('/api/admin/categories', {
+            method: 'GET',
+            headers: {
+              'Content-Type': 'application/json',
+            },
+          });
+  
+          const result = await res.json();
+          const data: CategoryData[] = result.categories;
+          console.log("APIレスポンス:", data); 
+          if (!res.ok) {
+            throw new Error();
+          } else {
+            setCategoryOptions(
+              data.map((cat) => ({
+                id: cat.id,
+                name: cat.name,
+              }))
+            );
+          }
+        
+        } catch (error) {
+          console.error('エラーが発生しました。', error);
+          setCategoryOptions([])
+        } finally {
+          setIsLoading(false);
+        }
+      }
+  
+      fetcher();
+    },[]);
+
   return (
     <form className='adminForm' id='myForm' onSubmit={handleSubmit(submitFunction)}>
       <label className='adminFormTitle' htmlFor='title'>タイトル</label>
